+++ sdk/resourcemanager/virtualmachineimagebuilder/armvirtualmachineimagebuilder/CHANGELOG.md
@@ -1,5 +1,35 @@
 # Release History
 
+## 1.1.0 (2022-05-30)
+### Features Added
+
+- New const `ProvisioningErrorCodeBadValidatorType`
+- New const `ProvisioningErrorCodeUnsupportedValidatorType`
+- New const `RunSubStateValidating`
+- New const `ProvisioningErrorCodeNoValidatorScript`
+- New const `ProvisioningErrorCodeBadStagingResourceGroup`
+- New function `ImageTemplatePowerShellValidator.MarshalJSON() ([]byte, error)`
+- New function `*ImageTemplatePropertiesValidate.UnmarshalJSON([]byte) error`
+- New function `*ImageTemplateShellValidator.UnmarshalJSON([]byte) error`
+- New function `*ImageTemplatePowerShellValidator.GetImageTemplateInVMValidator() *ImageTemplateInVMValidator`
+- New function `ImageTemplateShellValidator.MarshalJSON() ([]byte, error)`
+- New function `*ImageTemplateShellValidator.GetImageTemplateInVMValidator() *ImageTemplateInVMValidator`
+- New function `ImageTemplatePropertiesValidate.MarshalJSON() ([]byte, error)`
+- New function `*ImageTemplateInVMValidator.GetImageTemplateInVMValidator() *ImageTemplateInVMValidator`
+- New function `*ImageTemplatePowerShellValidator.UnmarshalJSON([]byte) error`
+- New struct `ImageTemplateInVMValidator`
+- New struct `ImageTemplatePowerShellValidator`
+- New struct `ImageTemplatePropertiesValidate`
+- New struct `ImageTemplateShellValidator`
+- New struct `ProxyResource`
+- New field `SystemData` in struct `TrackedResource`
+- New field `SystemData` in struct `RunOutput`
+- New field `ExactStagingResourceGroup` in struct `ImageTemplateProperties`
+- New field `StagingResourceGroup` in struct `ImageTemplateProperties`
+- New field `Validate` in struct `ImageTemplateProperties`
+- New field `SystemData` in struct `Resource`
+
+
 ## 1.0.0 (2022-05-18)
 ### Breaking Changes